requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
gunicorn==21.2.0
chatterbot==1.2.8
chatterbot-corpus==1.2.0